add item to shopping cart (amount is double) use case

+++ src/java/test/Bridge/Proxy.java
@@ -77,7 +77,7 @@ public class Proxy implements IService {
     }
 
     @Override
-    public Response addItemToShoppingCart(ItemFacade itemToInsert, int amount, String shopName, String visitorName) {
+    public Response addItemToShoppingCart(ItemFacade itemToInsert, double amount, String shopName, String visitorName) {
         return null;
     }
